Update README.md
Included in my files are my Design Thinking for my Rare Books project.  I have attached photos of my Empathise, Define, Ideate, and Prototype. Included in my sketches is an Admin and Employee log in screen. I also included photos of a  Admin system and inventory screen. In addition to that is an Employee system/ inventory screen that will be slightly different from the Admin screens.